Move away from using `terminfo` codes
They are meant for full-screen applications, not the command line.
Additionally, when a full-screen app exits, it sets the terminal back to
normal mode, but not all such exits cause line-init to be called. Thus,
when returning from a full-screen-ap, the command line can then
unexpectedly be in normal mode instead of app mode, causing certain
keyboard shortcuts to fail.