Added a section for usage with Oh My Zsh
As Oh My Zsh might as well be the most popular framework for managing zsh configs I thought an own section might be helpful. I tried the the plugin ```zsh-autocomplete``` which comes directly with Oh My ZshTook but like this one way more. It took me a bit as someone not yet that proficient with working in the shell to install and make it work and think the short sub-section might be helpful for folks like me.

(For your interest: This is my first pull request so please excuse me if I did not follow the conventions for Pull Requests correctly)